+++ src/lib.ts
@@ -4,7 +4,7 @@ export { ComplexDecimal } from './ComplexDecimal';
 export { MultiArray } from './MultiArray';
 export { CoreFunctions } from './CoreFunctions';
 export { LinearAlgebra } from './LinearAlgebra';
-export { MathObject } from './MathObject';
+export { MathOperation } from './MathOperation';
 export { Configuration } from './Configuration';
 export { SymbolTable } from './SymbolTable';
 export * as AST from './AST';